wip(frontend): move collective checkbox for all questions
